fix: avoid unneccesary temporary layout during build --oci
With #2935 in place we can now translate a Docker archive tarball into
an OCI-SIF without creating a redundant copy of the image as a temporary
OCI layout.

We've previously been processing the buildkit tarball as an oci-archive,
which still requires extraction to a temporary layout, as ggcr's tarball
package doesn't accept an archived layout without a manifest.json.

Because the buildkit output tarball includes the Docker specific
manifest.json, in addition to the standard OCI layout files, it can be
treated as a docker-archive - now avoiding the need for the temporary
layout.

Significantly reduces disk space needed by `build --oci`, and the
associated I/O.

Fixes #2308