oci: --bind / --mount for data containers
Allow OCI-SIF data container created with `singularity data package` to
be mounted using the `--bind` and `--mount` flags.

An `image-source=/` options is required on the bind to mount from the
root of the data container layer. A shortcut `--data` flag that does not
require this will be added in future.

Closes #2873