feat: push --layer-format tar
Add a `--layer-format` flag for the `singularity push` command. This
applies only to pushes of OCI-SIF images (including data containers).

When `--layer-format` is not set, or is set to `squashfs` then the
OCI-SIF image is pushed with layers in squashfs format. These can only
be used by Singularity.

When `--layer-format` is set to `tar` then the OCI-SIF image is pushed
with its layers converted on-the-fly to TAR format. The resulting pushed
image can be used by other OCI runtimes.

Closes #2848