feat: --data shortcut for --bind with image-src=/
Provide a `--data` flag that accepts a `<src image>:<dst>` pairing, and
has the same effect as `--bind <src image>:<dst>:image-src=/`.

This allow a data container created using the `singularity data` command
to be used with a matching, and intuitive, `--data` action flag. It avoids the user
having to know that they must bind with the additional `image-src=/`
option.

Closes #2979