feat: mount OCI-SIF overlays and support --writable
When an OCI-SIF contains an ext3 overlay layer, created with
`singularity overlay create`, it is now mounted when the container is
run.

By default, an embedded overlay is mounted read-only. In OCI-Mode we
have a `--writable-tmpfs` in place by default, so changes can still be
made to the container at runtime, but they are made in the ephemeral
tmpfs, and do not write into the overlay.

To write changes into the embedded overlay, the container must be
started with the `--writable` flag, which is now supported in OCI-Mode.

Embedded overlay layers are handled by extending the code that
previously dealt with user specified `--overlay`s.

Note that fuse2fs >=1.46.6 is required to use an embedded overlay, as
older versions do not support the `-o` (offset) flag.

Closes #2868
Closes #2869