fix: don't tmp-sandbox fallback for explicit OCI-SIF
When a user asks to run a `docker://` or other URI in `--oci` mode, then
we try to pull it to an OCI-SIF, but we can fall back to a tmp-sandbox
rootfs if this fails. This fall-back uses the pkg/ocibundle/native code,
which creates a bundle with a directory rootfs from an OCI source.

We don't currently have any code to create a tmp-sandbox rootfs out of an
OCI-SIF file. If a user is asking to run an OCI-SIF file, rather than an
OCI URI then we should not attempt a fall-back if

Additionally, fix the `ocibundle.Delete` function so that it will delete
a partial bundle (that has not been fully assembled), rather than
raising a confusing error and leaving remnants on the disk.

Fixes #3099