Fix SquashFS extraction when it contains directory symlinks
A SquashFS image might contain directory symlinks such as `/var/tmp -> /tmp`.
Before extraction a base environment is created (`makeBaseEnv`) by
pre-creating some directories. Among them `/var/tmp`.
Although unsquashfs is called with `-force` it fails to replace that by
the symlink in the image as it removes only directories.

Introduce a pre-extraction step listing the contents of the image and
searching for all entries which are existing directories. Remove those
to avoid the issue.

Fixes #3084