feat: oci-sif overlay sync
When a writable overlay is added to an OCI-SIF image, the digests for
the ext3 layer in the OCI structures are correct.

When the overlay is written to, its true digest no long matches those
listed in the OCI structures, and recorded against the descriptor in the
OCI-SIF.

The new `singularity overlay sync` command will update the digests for
an overlay, in an OCI-SIF file, that has been modified. It does this by:

* Modifying the OCI.Blob digest recorded for the SIF descriptor.
* Updating and re-writing the OCI image and index structures in the
  OCI-SIF.

At this time, this sync task is only invoked by running `singularity
overlay sync` manually. Future PRs may invoke the sync automatically
prior to pushing the image to a remote destination.

Closes #3119