oci: pull OCI-SIF with overlay with --keep-layers
When we pull an OCI-SIF with `--keep-layers` (so that layers are not
squashed), ensure that we can pull an image that contains an overlay.

This requires us to check whether the last layer is ext3. If it is, it
is an overlay and we write it to the OCI-SIF as-is.

Fixes #3136