feat: pull OCI-SIF with overlay, without --keep-layers
When we are pulling an OCI image to an OCI-SIF then by default we squash
the image layers down to a single layer for the OCI-SIF file.

This doesn't make sense when the last layer is an overlay - users will
generally expect an image pushed with a writable overlay remains
writable when pulled.

When pulling to OCI-SIF without `--keep-layers` squash all the squashfs
layers, but preserve a separate final overlay layer.

Fixes #3135