fix: enforce --platform for all OCI images
Since 4.x we have had `--platform` / `--arch` support on pull from OCI
sources. The `--platform` defaults to the host native platform.

In native mode, we were entirely relying on ggcr code to apply filtering
/ limits by platform. ggcr only does this when handling an image via an
image index, not a manifest directly. Therefore, it was possible for a
pull images (that do not have an image index), with an architecture not
expected.

In OCI-Mode a platform check prevented this, but occured very late...
after layers were copied.

Move the main platform checks in to the `ociimage.LocalImage` /
`ociimage.LocalImageLayout` functions, so that they are as early as
possible (before layers are retrieved), and function the same in native
and OCI mode.

When an image is retrieved from the cache - because the manifest/index
digest in the registry matches a cache entry - check the architecture /
platform.

Adjust some error messages, while we are here, that refer to fetching
into the cache even when we aren't.

Ensure there is an e2e test that tries to pull from a single-arch
repository with a valid, but non-matching platform... rather than just
an invalid platform. The e2e test ordering ensures both the cached and
non-cached flows are tested.

Fixes #3158