fix: ensure source baseenv not overwritten in local build
When building from a local image, we need to insert the base environment
(default runscripts etc), into the rootfs after extracting it. This is
because of an issue with `unsquashfs` that causes it to fail in the presence
of certain symlinks (#3151).

We also want any runscript etc. in the source image to be preservered,
and not replaced with defaults.

Modify the base environment insertion to accept a bool controlling
whether it overwrites any existing files. For the local image case,
where we insert the baseenv after extraction, do not overwrite. In all
other cases keep existing behaviour by setting overwrite to true.

Fixes #3353