fix: improve build --oci error for missing XDG_RUNTIME_DIR
The buildkit code that we call as a dependency uses XDG_RUNTIME_DIR
in multiple locations, so we cannot handle it being unset.

Add a specific error for rootless build --oci with XDG_RUNTIME_DIR
unset, so it is clear that this env var must be set.

I don't think we should be setting XDG_RUNTIME_DIR ourselves, to a
tmpdir we create. There is a strong assumption that this value is a
genuine session specific directory that is configured by systemd, or
by the user manually.

Fixes #3410