fix: use username, not name for conmon/singularity stateDir
For `singularity oci` commands (not `--oci` mode), a conmon/singularity
state dir is required. The path was erroneously being computed based on
the user's Name, instead of Username.

This shows as a test failure on EL10 where the Name != Username for
`root / Super User`, as opposed to other distributions that use `root /
root`.

Fixes #3456