fix: ensure network e2e tests run
Network e2e tests were failing to run due to an invalid container name
(with newer CNI plugins / distro) in the `require.Network(t)` function.

Fix the container name so that the requirement function passes, and the
tests will run properly.

While we are in this code, ensure some other network tests run with
named subtests for convenience when reviewing the test logs.

Fixes #3433