feat: pull -oci --with-cosign support
When using `singularity pull --oci` to pull an image from an OCI
registry (or registry backed library) into an OCI-SIF it is now
possible to pull any cosign signatures associated with the image using
a new `--with-cosign` flag.

Pulling cosign signatures requires that the image in the registry is
in a format where it will not be mutated as it is pulled into the
OCI-SIF. Any mutations change the image manifest, which would
invalidate the signatures. Therefore:

* The image must have SquashFS layers - i.e. it was created and signed
with Singularity, and is not a standard tar layer Docker / OCI image.
* The '--keep-layers' flag must be used if the image has more than one
layer.

There are important limitations to signature pulling:

* If an image & signatures are updated at the registry during a pull
`--with-cosign` by tag, it is possible that a mistmatched signature is
retrieved. Pulling by digest avoids this issue.
* Pulling signatures disables OCI-SIF caching.

The caching issue is difficult to overcome, as Singularity's caching
of OCI -> SIF pulls is based around the assumption that a SIF contains
a single OCI image and the image manifest digest can be used as the
cache key. An OCI-SIF containing cosign signatures now holds
multiple (container and signature) images. We cannot cache based on
the container manifest digest alone. We cannot add a simple `-signed`
suffix to the cache key, as the signatures associated with an image at
the registry can change independently of the container image itself.

In the mid to long-term it would be beneficial to rewrite the OCI pull
/ push handling to use the SourceSink interfaces and implementations
in sylabs/oci-tools, which accommodate multi-image SIFs more
directly. This would require a complete re-think of caching in
Singularity - which is overdue but dependent on a major version bump.