chore: disable usetesting linter on e2e/ code
The usetesting linter shows a lot of lint on e2e code for tempdir
handling. The e2e-tests are written, and have helper functions,
designed around keeping temp directories inside a single per-run outer
directory, often with somewhat useful naming. This would be broken by
a switch to t.TempDir.

The usetesting linter shows a lot of lint on e2e code for env var
handling. It's not often simple to switch to t.Setenv - the scope of
env vars is more complex due to the e2e.Singularitycmd construct etc.