e2e: basic nested in docker/podman testing
Basic testing of minimal container build and execution nested under
privileged docker, and rootless podman.

The docker/podman containerized singularity is built from the source
tree in which the e2e-tests are running.

We verify basic nested exec of a container with/without userns (-u),
an OCI-Mode container (always uses a userns), and simple build to
native SIF.

Manual testing performed on Fedora 41, with:

    Docker version 28.0.0, build f9ced58
    podman version 5.3.2