deps: drop conmon submodule
Do not include / build conmon sources with Singularity.

conmon is only required for `singularity oci` commands, which are
limited and rarely used. It is not required for the `--oci` mode.

A suitable version of `conmon` is now available in all EL versions, and
in newer Ubuntu versions. For other distributions it is a trivial single
binary install from GitHub. It is not necessary to include it here.