fix: don't set mode option on workdir bind mount
runc v1.2+ checks that a bind mount definition doesn't have any
fs-specific options set on it, as these will not be applied.

We are erroneously setting `mode-777` on a `--workdir` bind mount prior
to this PR, so `--workdir` causes a failure where runc v1.2+ is
installed, and crun is not installed.

Remove the incorrect option. It likely came from a copy/paste of a
`tmpfs` definition, and has never been correct/effective here.

Fixes #3716