fix: fail if --security options can't be applied
When a `--security` option is supplied on the command line, it is
reasonable to expect the limitations requested will be applied to the
container.

Previously, when a limitation was not applied due to a subsystem (e.g.
SELinux) being unavailable, a warning was displayed but execution
continued. The warning is easily overlooked or not observed when running
via scripts / automation.

Modify the flow so that we now exit with a fatal error when a `--security`
option cannot be applied.