fix: mount selinuxfs for setuid with --security:selinux
When `--security:selinux:<context>` is requested, the
`opencontainers/selinux.GetEnabled()` function is run in the container
mount namespace to ascertain whether selinux is enabled. This function
needs to access `/sys/fs/selinux`.

Prior to this PR, `/sys/fs/selinux` is only mounted in user-namespace
flows, where `/sys` is a recursive bind mount.

To fix selinux detection, mount `/sys/fs/selinux` as an selinuxfs when
a user namespace is not being used.